refactor(fe): add cocktailbar api functions
